Agents - remove "Selection" and "Go" main menu items
